Refactor server to serve static files and update deploy configs
