+++ client/.env.example
@@ -1,0 +1,2 @@
+VITE_API_URL=https://your-backend-service.onrender.com
+
