+++ package.json
@@ -7,6 +7,8 @@
     "dev": "NODE_ENV=development tsx server/index.ts",
     "build": "vite build && esbuild server/index.ts --platform=node --packages=external --bundle --format=esm --outdir=dist",
     "start": "NODE_ENV=production node dist/index.js",
+    "server:build": "esbuild server/index.ts --platform=node --packages=external --bundle --format=esm --outdir=dist",
+    "server:start": "NODE_ENV=production node dist/index.js",
     "check": "tsc",
     "db:push": "drizzle-kit push",
     "frontend:build": "vite build"
