+++ types/shims.d.ts
@@ -1,0 +1,2 @@
+declare module "cors";
+
